revert: 드래그&드랍/복붙 훅 제외
